% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dblr.R
\name{dblr}
\alias{dblr}
\title{function to double things}
\usage{
dblr(x)
}
\arguments{
\item{x}{a number}
}
\value{
the number doubled
}
\description{
This function takes a number and returns it doubled.
}
\examples{
dblr(2)
dblr(3)
}
