% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubler-package.R
\docType{package}
\name{doubler-package}
\alias{doubler}
\alias{doubler-package}
\title{doubler: Doubles everything you give it}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Bilguun Boldbaatar \email{bilguunboldbaatar@19.com}

Authors:
\itemize{
  \item Bat Tulga \email{battulga@example.com}
}

}
\keyword{internal}
