% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubler.R
\name{doubler}
\alias{doubler}
\title{Double Values in R}
\usage{
doubler(value)
}
\arguments{
\item{value}{A numeric value to be doubled.}
}
\description{
Doubler is a testing package aimed at GitHub Actions testing. The function is usable, but minimal and uneccessary in application.
}
