% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dblr.R
\name{dblr}
\alias{dblr}
\title{Function to double things}
\usage{
dblr(x)
}
\arguments{
\item{x}{A number to double}
}
\value{
The number doubled
}
\description{
This function takes a number and doubles it.
}
\examples{
dblr(4)
}
